requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
Pillow>=8.0.0
psutil>=5.8.0
pytest>=6.0.0
matplotlib>=3.3.0
